Move Dater Button to Atoms
