+++ app/components/map-panel/map-panel-component.js
@@ -13,7 +13,7 @@ import { connect, Dispatch } from 'react-redux';
 import MapPanelStyles from './map-panel-styles';
 
 import { H2, Caption2 } from '../../components/ui-kit/typography';
-import DaterButton from '../../components/ui-kit/dater-button';
+import DaterButton from '../../components/ui-kit/atoms/dater-button';
 import MapPanelSelfieUploading from './map-panel-selfie-uploading';
 import MapPanelSelfieUploadedByMe from './map-panel-selfie-uploaded-by-me';
 import MapPanelSelfieUploadedByTarget from './map-panel-selfie-uploaded-by-target';
