+++ app/screens/floating-screen.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { StyleSheet } from 'react-native';
 
 import DaterModal from '../components/ui-kit/dater-modal';
-import DaterButton from '../components/ui-kit/dater-button';
+import DaterButton from '../components/ui-kit/atoms/dater-button';
 
 type Props = {
   navigation: any,
