+++ app/screens/login/login-screen.js
@@ -7,7 +7,7 @@ import {
 import LinearGradient from 'react-native-linear-gradient';
 
 import DaterModal from '../../components/ui-kit/dater-modal';
-import DaterButton from '../../components/ui-kit/dater-button';
+import DaterButton from '../../components/ui-kit/atoms/dater-button';
 import { H1, Body, Caption2 } from '../../components/ui-kit/typography';
 import { width } from '../../constants';
