+++ app/screens/profile/name-screen.js
@@ -7,7 +7,7 @@ import {
 
 import DaterTextInput from '../../components/ui-kit/atoms/dater-text-input';
 import DaterModal from '../../components/ui-kit/dater-modal';
-import DaterButton from '../../components/ui-kit/dater-button';
+import DaterButton from '../../components/ui-kit/atoms/dater-button';
 import { H2 } from '../../components/ui-kit/typography';
 
 const nameIcon = require('../../assets/icons/name/name.png');
