+++ app/screens/profile/upload-photo-screen.js
@@ -5,7 +5,7 @@ import {
 } from 'react-native';
 
 import DaterModal from '../../components/ui-kit/dater-modal';
-import DaterButton from '../../components/ui-kit/dater-button';
+import DaterButton from '../../components/ui-kit/atoms/dater-button';
 import { H2, Body } from '../../components/ui-kit/typography';
 
 const uploadPhotoIcon = require('../../assets/icons/upload-photo/upload-photo.png');
